+++ library/core/datastore/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/core/datastore/di/DataStoreModule.kt
@@ -15,7 +15,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.mihaicristiancondrea.android.libs.apptoolkit.core.datastore.data.local.di
+package com.mihaicristiancondrea.android.libs.apptoolkit.core.datastore.di
 
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.common.data.local.CommonDataStoreCore
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.datastore.data.local.CommonDataStore